+++ docs/conf.py
@@ -47,7 +47,7 @@ extensions = [
     'sphinx.ext.autosummary',
     'sphinx.ext.doctest',
     'sphinx.ext.inheritance_diagram',
-    'numpydoc',
+    #   'numpydoc',
 ]
 
 # Add any paths that contain templates here, relative to this directory.
@@ -131,12 +131,17 @@ numpydoc_show_class_members = False
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = 'alabaster'
+#html_theme = 'alabaster'
+#html_theme = 'nature'
+html_theme= 'sphinx_rtd_theme'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-# html_theme_options = {}
+html_theme_options = {
+ #   "rightsidebar": "false",
+ #   "relbarbgcolor": "black"
+}
 
 # Add any paths that contain custom themes here, relative to this directory.
 # html_theme_path = []